refactor: Simplify Bybit execution event parsing and improve type safety
- Changed `order_id` type from `size_t` to `uint64_t` in `ExecutionEventInterface` for better type safety
- Simplified `BybitExecutionEventParser` by:
  - Removing template complexity and `MultiPoolHolder` dependency
  - Using direct memory pools for each event type
  - Adding comprehensive JSON parsing with error checking
  - Integrating `TradingPairFactory` for symbol resolution
- Removed commented-out template code and unused `MultiPoolHolder` implementation
- Updated example20 to demonstrate new parser functionality:
  - Added simdjson parsing example
  - Integrated with `TradingPairFactory`
  - Cleaned up test code
- Added simdjson dependency to relevant CMakeLists
- Improved error handling throughout the parsing process

Key improvements:
1. More robust JSON parsing with proper error checking
2. Simplified memory management with direct pools
3. Better type safety with uint64_t for order IDs
4. Reduced template complexity
5. Added trading pair resolution capability

+++ aoe/bybit/execution_event/i_types.h
@@ -15,7 +15,7 @@ class ExecutionEventInterface
     virtual double ExecQty() const = 0;
     virtual double ExecQtyAbs() const { return exec_qty_; }
     virtual double ExecValue() const { return exec_value_; }
-    virtual size_t OrderId() const { return order_id_; };
+    virtual uint64_t OrderId() const { return order_id_; };
     virtual common::ExchangeId ExchangeId() const { return exchange_id; };
     virtual common::TradingPair TradingPair() const { return trading_pair_; };
 
@@ -23,7 +23,7 @@ class ExecutionEventInterface
     virtual void SetExecPrice(double exec_price) { exec_price_ = exec_price; };
     virtual void SetExecQty(double exec_qty) { exec_qty_ = exec_qty; };
     virtual void SetExecValue(double exec_value) { exec_value_ = exec_value; };
-    virtual void SetOrderId(size_t order_id) { order_id_ = order_id; };
+    virtual void SetOrderId(uint64_t order_id) { order_id_ = order_id; };
     virtual void SetTradingPair(common::TradingPair trading_pair) {
         trading_pair_ = trading_pair;
     };
@@ -37,7 +37,7 @@ class ExecutionEventInterface
     double exec_price_             = 0;
     double exec_qty_               = 0;
     double exec_value_             = 0;
-    size_t order_id_               = 0;
+    uint64_t order_id_             = 0;
     common::ExchangeId exchange_id = common::ExchangeId::kBybit;
     common::TradingPair trading_pair_;
 };
